Added an option that allows a JSON object to be used to dictate the sequence of xml elements when converting a json object to xml
A JSON object, that describes the order of elements on a type, can now be passed into easy xml via configuration. When calling the 'render' function, an additional argument that states the 'Type' of object being passed is now available. The render function can then use the supplied 'Type' along with the schema to ensure that the elements on the resulting xml are in the correct order.